Rearrange library platform utils
