+++ packages/library/src/androidMain/kotlin/io/realm/internal/platform/LogCatLogger.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.realm.internal
+package io.realm.internal.platform
 
 import android.util.Log
 import io.realm.log.LogLevel
