+++ packages/library/src/androidMain/kotlin/io/realm/internal/platform/RealmInitializer.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.realm.internal
+package io.realm.internal.platform
 
 import android.content.Context
 import androidx.startup.Initializer
