+++ packages/library/src/commonMain/kotlin/io/realm/Realm.kt
@@ -18,8 +18,9 @@ package io.realm
 import io.realm.internal.RealmReference
 import io.realm.internal.SuspendableNotifier
 import io.realm.internal.SuspendableWriter
-import io.realm.internal.WeakReference
-import io.realm.internal.runBlocking
+import io.realm.internal.platform.WeakReference
+import io.realm.internal.platform.runBlocking
+import io.realm.internal.platform.threadId
 import io.realm.interop.NativePointer
 import io.realm.interop.RealmInterop
 import kotlinx.atomicfu.AtomicRef
