+++ packages/library/src/commonMain/kotlin/io/realm/internal/platform/WeakReference.kt
@@ -1,0 +1,11 @@
+package io.realm.internal.platform
+
+/**
+ * Platform agnostic _WeakReference_ wrapper.
+ *
+ * Basically just wrapping underlying platform implementations.
+ */
+expect class WeakReference<T : Any>(referred: T) {
+    fun clear()
+    fun get(): T?
+}
