+++ packages/library/src/darwinCommon/kotlin/io/realm/internal/platform/CoroutineUtils.kt
@@ -1,0 +1,18 @@
+package io.realm.internal.platform
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.newSingleThreadContext
+import kotlin.coroutines.CoroutineContext
+
+// Expose platform runBlocking through common interface
+public actual fun <T> runBlocking(context: CoroutineContext, block: suspend CoroutineScope.() -> T): T {
+    return kotlinx.coroutines.runBlocking(context, block)
+}
+
+/**
+ * The default dispatcher for Darwin platforms spawns a new thread with a run loop.
+ */
+actual fun singleThreadDispatcher(id: String): CoroutineDispatcher {
+    return newSingleThreadContext(id)
+}
