+++ packages/library/src/darwinCommon/kotlin/io/realm/internal/platform/NSLogLogger.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.realm.internal
+package io.realm.internal.platform
 
 import io.realm.log.LogLevel
 import io.realm.log.RealmLogger
