+++ packages/library/src/jvm/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -1,0 +1,12 @@
+package io.realm.internal.platform
+
+actual fun threadId(): ULong {
+    return Thread.currentThread().id.toULong()
+}
+
+actual fun <T> T.freeze(): T = this
+
+actual val <T> T.isFrozen: Boolean
+    get() = false
+
+actual fun Any.ensureNeverFrozen() {}
