+++ packages/library/src/jvmMain/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -1,0 +1,10 @@
+package io.realm.internal.platform
+
+import io.realm.log.RealmLogger
+
+@Suppress("FunctionOnlyReturningConstant")
+actual fun appFilesDirectory(): String {
+    return "."
+}
+
+actual fun createDefaultSystemLogger(tag: String): RealmLogger = StdOutLogger(tag)
