+++ packages/library/src/macosMain/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -1,0 +1,6 @@
+package io.realm.internal.platform
+
+@Suppress("FunctionOnlyReturningConstant")
+actual fun appFilesDirectory(): String {
+    return "."
+}
