+++ test/src/androidTest/kotlin/io/realm/InstrumentedTests.kt
@@ -19,7 +19,7 @@ package io.realm
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import io.realm.internal.RealmInitializer
+import io.realm.internal.platform.RealmInitializer
 import io.realm.util.PlatformUtils
 import io.realm.util.Utils.createRandomString
 import org.junit.After
